[Refactoring ⚙️] react-docgen을 이용해 스토리북의 argTypes 자동 추론
