+++ src/components/Button/RoundButton.stories.tsx
@@ -1,33 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
 import RoundButton from "@components/Button/RoundButton"
-import { ICON_MAP } from "@components/Icon/Icon"
 
 const meta: Meta<typeof RoundButton> = {
   component: RoundButton,
   title: "components/RoundButton",
   tags: ["autodocs"],
   parameters: { layout: "centered" },
-  argTypes: {
-    leftIcon: {
-      control: {
-        type: "select",
-      },
-      options: Object.keys(ICON_MAP),
-    },
-    rightIcon: {
-      control: {
-        type: "select",
-      },
-      options: Object.keys(ICON_MAP),
-    },
-    size: {
-      control: {
-        type: "radio",
-      },
-      options: ["lg", "sm"],
-    },
-  },
 }
 
 export default meta
