+++ src/components/ImgCheckBox/StyledImgCheckBox.ts
