+++ src/pages/Recommend/Machine/List.tsx
@@ -1,7 +1,7 @@
 import { ForwardedRef } from "react"
 
 import Avatar from "@components/Avatar/Avatar"
-import ImgCheckBox from "@components/CheckBox/ImgCheckBox"
+import ImgCheckBox from "@components/ImgCheckBox/ImgCheckBox"
 import SpeechBubble from "@components/SpeechBubble/SpeechBubble"
 
 import { MachineList } from "@typpes/type"
